requirements.txt
streamlit
youtube-transcript-api
langchain
faiss-cpu
huggingface-hub
sentence-transformers
python-dotenv
langchain-groq